+++ src/tables/cmap/format10.rs
@@ -16,17 +16,19 @@ pub fn parse(data: &[u8], code_point: u32) -> Option<u16> {
     glyphs.get(idx)
 }
 
-pub fn codepoints(data: &[u8], mut f: impl FnMut(u32)) {
+pub fn codepoints(data: &[u8], mut f: impl FnMut(u32)) -> Option<()> {
     let mut s = Stream::new(data);
     s.skip::<u16>(); // format
     s.skip::<u16>(); // reserved
     s.skip::<u32>(); // length
     s.skip::<u32>(); // language
-    let first_code_point: u32 = try_opt_or!(s.read(), ());
-    let count: u32 = try_opt_or!(s.read(), ());
+    let first_code_point: u32 = s.read()?;
+    let count: u32 = s.read()?;
 
     for i in 0..count {
-        let code_point = try_opt_or!(first_code_point.checked_add(i), ());
+        let code_point = first_code_point.checked_add(i)?;
         f(code_point);
     }
+
+    Some(())
 }