+++ src/tables/cmap/format6.rs
@@ -20,16 +20,18 @@ pub fn parse(data: &[u8], code_point: u32) -> Option<u16> {
     glyphs.get(idx)
 }
 
-pub fn codepoints(data: &[u8], mut f: impl FnMut(u32)) {
+pub fn codepoints(data: &[u8], mut f: impl FnMut(u32)) -> Option<()> {
     let mut s = Stream::new(data);
     s.skip::<u16>(); // format
     s.skip::<u16>(); // length
     s.skip::<u16>(); // language
-    let first_code_point: u16 = try_opt_or!(s.read(), ());
-    let count: u16 = try_opt_or!(s.read(), ());
+    let first_code_point: u16 = s.read()?;
+    let count: u16 = s.read()?;
 
     for i in 0..count {
-        let code_point = try_opt_or!(first_code_point.checked_add(i), ());
+        let code_point = first_code_point.checked_add(i)?;
         f(u32::from(code_point));
     }
+
+    Some(())
 }