+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-- `Subtable::codepoints`
+- `cmap::Subtable::codepoints`
 
 ## [0.8.1] - 2020-07-29
 ### Added